split engine compiler
